Rename DoomObjectContainer file for consistency
