+++ common/m_doomobjcontainer.h
