+++ common/mobjinfo.cpp
@@ -2,7 +2,7 @@
 
 #include "mobjinfo.h"
 #include "info.h"
-#include "doom_obj_container.h"
+#include "m_doomobjcontainer.h"
 
 void D_Initialize_Mobjinfo(mobjinfo_t* source, int count)
 {
